// src/pages/smartReplenishment/pages/invoicesReview/columns.d.ts
// 请确保此处与真实字段完全匹配
export enum keys {
  inCode = 'inCode',                // 单据编号
  storeName = 'storeName',          // 门店名称
  inType = 'inType',                // 单据类型
  submitter = 'submitter',          // 提交人
  submitDateTime = 'submitDateTime',// 提交时间
}